+++ src/Crypto/WebAuthn/AttestationStatementFormat/Packed.hs
@@ -46,7 +46,8 @@ instance Show Format where
 data Statement = Statement
   { alg :: Cose.CoseSignAlg,
     sig :: Cose.CoseSignature,
-    x5c :: Maybe (NE.NonEmpty X509.SignedCertificate, IdFidoGenCeAAGUID)
+    -- The AAGUID extension is optional
+    x5c :: Maybe (NE.NonEmpty X509.SignedCertificate, Maybe IdFidoGenCeAAGUID)
   }
   deriving (Eq, Show)
 
@@ -97,35 +98,41 @@ instance M.AttestationStatementFormat Format where
 
   asfDecode _ xs =
     case (xs !? "alg", xs !? "sig", xs !? "x5c") of
-      (Just (CBOR.TInt algId), Just (CBOR.TBytes (Cose.CoseSignature -> sig)), Just (CBOR.TList x5cRaw)) -> do
+      (Just (CBOR.TInt algId), Just (CBOR.TBytes (Cose.CoseSignature -> sig)), mx5c) -> do
         alg <- Cose.toCoseSignAlg algId
-        x5c <- case NE.nonEmpty x5cRaw of
+        x5c <- case mx5c of
           Nothing -> pure Nothing
-          Just x5cBytes -> do
-            x5c@(signedCert :| _) <- forM x5cBytes $ \case
-              CBOR.TBytes certBytes ->
-                first (("Failed to decode signed certificate: " <>) . Text.pack) (X509.decodeSignedCertificate certBytes)
-              cert ->
-                Left $ "Certificate CBOR value is not bytes: " <> Text.pack (show cert)
+          Just (CBOR.TList x5cRaw) -> case NE.nonEmpty x5cRaw of
+            Nothing -> pure Nothing
+            Just x5cBytes -> do
+              x5c@(signedCert :| _) <- forM x5cBytes $ \case
+                CBOR.TBytes certBytes ->
+                  first (("Failed to decode signed certificate: " <>) . Text.pack) (X509.decodeSignedCertificate certBytes)
+                cert ->
+                  Left $ "Certificate CBOR value is not bytes: " <> Text.pack (show cert)
 
-            let cert = X509.getCertificate signedCert
-            aaguidExt <- case X509.extensionGetE (X509.certExtensions cert) of
-              Just (Right ext) -> pure ext
-              Just (Left err) -> Left $ "Failed to decode certificate aaguid extension: " <> Text.pack err
-              Nothing -> Left "Certificate aaguid extension is missing"
-            pure $ Just (x5c, aaguidExt)
+              let cert = X509.getCertificate signedCert
+              aaguidExt <- case X509.extensionGetE (X509.certExtensions cert) of
+                Just (Right ext) -> pure $ Just ext
+                Just (Left err) -> Left $ "Failed to decode certificate aaguid extension: " <> Text.pack err
+                Nothing -> pure Nothing
+              pure $ Just (x5c, aaguidExt)
+          Just _ -> Left $ "CBOR map didn't have expected value types (alg: int, sig: bytes, [optional] x5c: non-empty list): " <> Text.pack (show xs)
         pure $ Statement {..}
-      _ -> Left $ "CBOR map didn't have expected value types (alg: int, sig: bytes, x5c: list): " <> Text.pack (show xs)
+      _ -> Left $ "CBOR map didn't have expected value types (alg: int, sig: bytes, [optional] x5c: non-empty list): " <> Text.pack (show xs)
 
   asfEncode _ Statement {..} =
-    let encodedx5c = case x5c of
-          Nothing -> []
-          Just (certChain, _) -> map (CBOR.TBytes . X509.encodeSignedObject) $ toList certChain
-     in CBOR.TMap
-          [ (CBOR.TString "sig", CBOR.TBytes $ Cose.unCoseSignature sig),
-            (CBOR.TString "alg", CBOR.TInt $ Cose.fromCoseSignAlg alg),
-            (CBOR.TString "x5c", CBOR.TList encodedx5c)
-          ]
+    CBOR.TMap
+      ( [ (CBOR.TString "sig", CBOR.TBytes $ Cose.unCoseSignature sig),
+          (CBOR.TString "alg", CBOR.TInt $ Cose.fromCoseSignAlg alg)
+        ]
+          ++ case x5c of
+            Nothing -> []
+            Just (certChain, _) ->
+              let encodedx5c = map (CBOR.TBytes . X509.encodeSignedObject) $ toList certChain
+               in [ (CBOR.TString "x5c", CBOR.TList encodedx5c)
+                  ]
+      )
 
   type AttStmtVerificationError Format = VerificationError
 
@@ -153,7 +160,7 @@ instance M.AttestationStatementFormat Format where
           pure $ M.SomeAttestationType M.AttestationTypeSelf
 
         -- Basic, AttCA
-        Just (x5c@(certCred :| _), IdFidoGenCeAAGUID certAAGUID) -> do
+        Just (x5c@(certCred :| _), mbAAGUID) -> do
           let cert = X509.getCertificate certCred
               pubKey = X509.certPubKey cert
           -- Verify that sig is a valid signature over the concatenation of authenticatorData and clientDataHash using
@@ -176,8 +183,11 @@ instance M.AttestationStatementFormat Format where
 
           -- If attestnCert contains an extension with OID 1.3.6.1.4.1.45724.1.1.4 (id-fido-gen-ce-aaguid) verify that
           -- the value of this extension matches the aaguid in authenticatorData.
-          let aaguid = M.acdAaguid credData
-          unless (certAAGUID == aaguid) . failure $ CertificateAAGUIDMismatch certAAGUID aaguid
+          case mbAAGUID of
+            Just (IdFidoGenCeAAGUID certAAGUID) -> do
+              let aaguid = M.acdAaguid credData
+              unless (certAAGUID == aaguid) . failure $ CertificateAAGUIDMismatch certAAGUID aaguid
+            Nothing -> pure ()
 
           pure $
             M.SomeAttestationType $