+++ src/Crypto/WebAuthn/Metadata/Service/Processing.hs
@@ -60,9 +60,11 @@ import Data.FileEmbed (embedFile)
 import Data.HashMap.Strict (HashMap, (!?))
 import qualified Data.HashMap.Strict as HashMap
 import Data.Hourglass (DateTime)
+import Data.List.NonEmpty (NonEmpty, singleton)
 import qualified Data.List.NonEmpty as NE
 import Data.Text (Text)
 import qualified Data.Text as Text
+import Data.These (These (This))
 import qualified Data.X509 as X509
 import qualified Data.X509.CertificateStore as X509
 import qualified Data.X509.Validation as X509
@@ -182,12 +184,10 @@ jwtToJson blob rootCert now = runExcept $ do
 -- relying party the `Crypto.WebAuthn.Metadata.Service.Types.mpNextUpdate`
 -- and `Crypto.WebAuthn.Metadata.Service.Types.mpEntries` fields are most
 -- important.
-jsonToPayload :: HashMap Text Value -> Either Text Service.MetadataPayload
+jsonToPayload :: HashMap Text Value -> These (NonEmpty Text) Service.MetadataPayload
 jsonToPayload value = case Aeson.parseEither metadataPayloadParser value of
-  Left err -> Left $ Text.pack err
-  Right payload -> case decodeMetadataPayload payload of
-    Left err -> Left err
-    Right result -> pure result
+  Left err -> This (singleton $ Text.pack err)
+  Right payload -> decodeMetadataPayload payload
 
 metadataPayloadParser :: HashMap Text Aeson.Value -> Aeson.Parser ServiceIDL.MetadataBLOBPayload
 metadataPayloadParser hm = case (hm !? "legalHeader", hm !? "no", hm !? "nextUpdate", hm !? "entries") of