+++ tests/MetadataSpec.hs
@@ -3,6 +3,7 @@
 
 module MetadataSpec (spec) where
 
+import Crypto.WebAuthn.Metadata (metadataBlobToRegistry)
 import Crypto.WebAuthn.Metadata.Service.Processing (RootCertificate (RootCertificate), fidoAllianceRootCertificate, jsonToPayload, jwtToJson)
 import Crypto.WebAuthn.Metadata.Service.WebIDL (MetadataBLOBPayload, entries, legalHeader, nextUpdate, no)
 import Data.Aeson (Result (Success), ToJSON (toJSON), decodeFileStrict, fromJSON)
@@ -13,16 +14,17 @@ import Data.HashMap.Strict ((!), (!?))
 import qualified Data.PEM as PEM
 import qualified Data.Text as Text
 import Data.Text.Encoding (decodeUtf8)
+import Data.These (These (That, These, This))
 import qualified Data.X509 as X509
 import qualified Data.X509.CertificateStore as X509
-import System.Hourglass (dateCurrent)
+import Spec.Util (predeterminedDateTime)
 import Test.Hspec (SpecWith, describe, it, shouldBe, shouldSatisfy)
 import Test.Hspec.Expectations.Json (shouldBeUnorderedJson)
 
 golden :: FilePath -> SpecWith ()
 golden subdir = describe subdir $ do
   it "can verify and extract the blob payload" $ do
-    origin <- Text.unpack . Text.strip . decodeUtf8 <$> BS.readFile ("tests/golden-metadata/" <> subdir <> "/origin")
+    origin <- Text.unpack . Text.strip . decodeUtf8 <$> BS.readFile ("tests/golden-metadata/" <> subdir <> "/origin.txt")
 
     certBytes <- BS.readFile $ "tests/golden-metadata/" <> subdir <> "/root.crt"
     let Right [PEM.pemContent -> pem] = PEM.pemParseBS certBytes
@@ -30,8 +32,7 @@ golden subdir = describe subdir $ do
         store = X509.makeCertificateStore [cert]
 
     blobBytes <- BS.readFile $ "tests/golden-metadata/" <> subdir <> "/blob.jwt"
-    now <- dateCurrent
-    let Right result = jwtToJson blobBytes (RootCertificate store origin) now
+    let Right result = jwtToJson blobBytes (RootCertificate store origin) predeterminedDateTime
 
     Just expectedPayload <- decodeFileStrict $ "tests/golden-metadata/" <> subdir <> "/payload.json"
 
@@ -50,8 +51,9 @@ golden subdir = describe subdir $ do
   it "can decode and reencode the payload to the partially parsed JSON" $ do
     Just value <- decodeFileStrict $ "tests/golden-metadata/" <> subdir <> "/payload.json"
     case jsonToPayload value of
-      Left err -> fail $ show err
-      Right _result -> pure ()
+      This err -> fail $ show err
+      These err _result -> fail $ show err
+      That _result -> pure ()
 
 spec :: SpecWith ()
 spec = do
@@ -61,5 +63,11 @@ spec = do
   describe "fidoAllianceRootCertificate" $ do
     it "can validate the payload" $ do
       blobBytes <- BS.readFile "tests/golden-metadata/big/blob.jwt"
-      now <- dateCurrent
-      jwtToJson blobBytes fidoAllianceRootCertificate now `shouldSatisfy` isRight
+      jwtToJson blobBytes fidoAllianceRootCertificate predeterminedDateTime `shouldSatisfy` isRight
+  describe "MDS with errors" $ do
+    it "can process an MDS file with errors" $ do
+      blobBytes <- BS.readFile "tests/golden-metadata/big/blob-with-errors.jwt"
+      case metadataBlobToRegistry blobBytes predeterminedDateTime of
+        Right (These _errs _res) -> pure ()
+        Right _thisThat -> error "Expected parsing errors as well as registry"
+        Left err -> error $ Text.unpack err