+++ tests/Spec/Types.hs
@@ -119,11 +119,11 @@ instance Arbitrary (M.CollectedClientData c 'False) where
   arbitrary =
     M.CollectedClientData
       <$> arbitrary
-        <*> arbitrary
-        -- The crossOrigin value can't be roundtripped with Nothing values,
-        -- so let's just not generate Nothing values here
-        <*> (Just <$> arbitrary)
-        <*> arbitrary
+      <*> arbitrary
+      -- The crossOrigin value can't be roundtripped with Nothing values,
+      -- so let's just not generate Nothing values here
+      <*> (Just <$> arbitrary)
+      <*> arbitrary
 
 instance Arbitrary (M.AttestationObject 'False) where
   arbitrary = do
@@ -142,9 +142,9 @@ instance Arbitrary ArbitraryAttestationStatementFormat where
   arbitrary =
     elements
       [ ArbitraryAttestationStatementFormat None.Format
-      --ArbitraryAttestationStatementFormat Packed.Format,
-      --ArbitraryAttestationStatementFormat FidoU2F.Format,
-      --ArbitraryAttestationStatementFormat AndroidKey.Format
+      -- ArbitraryAttestationStatementFormat Packed.Format,
+      -- ArbitraryAttestationStatementFormat FidoU2F.Format,
+      -- ArbitraryAttestationStatementFormat AndroidKey.Format
       ]
 
 instance Arbitrary M.SignatureCounter where
@@ -237,7 +237,7 @@ instance Arbitrary M.AuthenticatorSelectionCriteria where
       <*> arbitrary
 
 instance Arbitrary M.AuthenticationExtensionsClientInputs where
-  arbitrary = pure M.AuthenticationExtensionsClientInputs
+  arbitrary = M.AuthenticationExtensionsClientInputs <$> arbitrary
 
 instance Arbitrary (M.CredentialOptions 'M.Registration) where
   arbitrary =
@@ -262,8 +262,11 @@ instance Arbitrary (M.CredentialOptions 'M.Authentication) where
       <*> arbitrary
       <*> arbitrary
 
+instance Arbitrary M.CredentialPropertiesOutput where
+  arbitrary = M.CredentialPropertiesOutput <$> arbitrary
+
 instance Arbitrary M.AuthenticationExtensionsClientOutputs where
-  arbitrary = pure M.AuthenticationExtensionsClientOutputs
+  arbitrary = M.AuthenticationExtensionsClientOutputs <$> arbitrary
 
 instance Arbitrary (M.Credential 'M.Registration 'False) where
   arbitrary =